edit: move code from app to Application
